Many updates:
configmaps: rewrite kwite://... schemes

code: separatesout the vast majority of the controller
implementation. The point is when / if a kubebuilder release requires a
re-scaffold of the project, this makes it easier to do the necessary
code moves into the new (re)scaffolding.

imagePullSecrets: fix bug that disallowed secrets

status: more useful reporting (and it actually works now).

Signed-off-by: Tom Hite <thite@vmware.com>